# docs/OpenAPI/openapi.yaml
openapi: 3.0.0
servers:
  - description: Api Back-end 
    url: http://localhost:8080
info:
  description: This
  version: "1.0.0"
  title: Direct-Sincro Api
tags:
  - name: Administradores
    description: Rotas de administrador
  - name: Subscritores
    description: Operações disponiveis para cidadaos que ja se subscreveram à api
  - name: Cidadao
    description: Operações disponiveis para cidadaos sem sessao iniciada
paths:
  /api/subscritores/{nif}/veiculos/{matricula}/eventos:
    post:
      tags:
        - Subscritores
      summary: Associar evento de transito ao proprietario
      operationId: "adicionarEventoAEmprestar"
      parameters:
      - name: nif
        in: path
        description: nif do cidadao
        required: true
        schema:
          type: integer
      - name: matricula
        in: path
        description: matricula do veiculo
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventoDeTransito"
      responses:
        '201':
          description: O pedido foi enviado com sucesso 
        '400':
          description: Algum parametro estava incorreto
        '409': 
          description: O pedido ja se encontra criado.

  /api/subscritores/{nif}/alugados/{matricula}/eventos:
    post:
      tags:
        - Subscritores
      summary: Associar evento de transito ao responsavel pelo Emprestimo
      operationId: "adicionarEventoEmprestado"
      parameters:
      - name: nif
        in: path
        description: nif do cidadao
        required: true
        schema:
          type: integer
      - name: matricula
        in: path
        description: matricula do veiculo
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventoDeTransito"
      responses:
        '201':
          description: O pedido foi enviado com sucesso 
        '400':
          description: Algum parametro estava incorreto
        '409': 
          description: O pedido ja se encontra criado.

  /api/subscritores/{nif}/emprestimos?matricula=string&datainicio=long:
    get:
      tags:
        - Subscritores
      summary: Requerer os Emprestimos deste veiculo com data de inicio pretendida
      operationId: "pedirEmprestimosComData"
      parameters:
      - name: nif
        in: path
        description: nif do cidadao
        required: true
        schema:
          type: integer
      - name: matricula
        in: query
        description: matricula do veiculo
        required: true
        schema:
          type: string
      - in: query
        name: datainicio
        required: true
        schema:
          type: string
        description: data em que Emprestimo iniciou
      responses:
        '200':
          description: Emprestimos do veiculo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Emprestimo'
        '400':
          description: Algum parametro estava incorreto
        '404': 
          description: O pedido nao foi encontrado

  /api/subscritores/{nif}/veiculos/{matricula}/emprestimos:
    get:
      tags:
        - Subscritores
      summary: Requerer os Emprestimos deste veiculo
      operationId: "pedirEmprestimos"
      parameters:
      - name: nif
        in: path
        description: nif do cidadao
        required: true
        schema:
          type: integer
      - name: matricula
        in: path
        description: matricula do veiculo
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Emprestimos do veiculo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Emprestimos'
        '400':
          description: Algum parametro estava incorreto
        '404': 
          description: O pedido nao foi encontrado
    post:
      tags:
        - Subscritores
      summary: Adicionar emprestimo ao veiculo
      operationId: "fazerEmprestimo"
      parameters:
      - name: nif
        in: path
        description: nif do cidadao
        required: true
        schema:
          type: integer
      - name: matricula
        in: path
        description: matricula do veiculo
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Emprestimo"
      responses:
        '201':
          description: O pedido foi enviado com sucesso 
        '400':
          description: Algum parametro estava incorreto
        '409': 
          description: O pedido ja se encontra criado.

  /api/subscritores/{nif}/veiculos/{matricula}/eventos/{numeroauto}/pedido:
    post:
      tags:
        - Subscritores
      summary: Fazer pedido de defesa
      description: Fazer pedido de defesa associado ao evento especificado
      operationId: "fazerPedido"
      parameters:
      - name: nif
        in: path
        description: nif do cidadao
        required: true
        schema:
          type: integer
      - name: matricula
        in: path
        description: matricula do veiculo
        required: true
        schema:
          type: string
      - name: numeroauto
        in: path
        description: numero do Auto
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PedidoDeDefesa"
      responses:
        '201':
          description: O pedido foi enviado com sucesso 
        '400':
          description: Algum parametro estava incorreto
        '409': 
          description: O pedido ja se encontra criado.
    get:
      tags:
        - Subscritores
      summary: Requerer o pedido de defesa
      operationId: "pedirPedido"
      parameters:
      - name: nif
        in: path
        description: nif do cidadao
        required: true
        schema:
          type: integer
      - name: matricula
        in: path
        description: matricula do veiculo
        required: true
        schema:
          type: string
      - name: numeroauto
        in: path
        description: numero do Auto
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Pedido de defesa feito para numero de auto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PedidoDeDefesa'
        '400':
          description: Algum parametro estava incorreto
        '404': 
          description: O pedido nao foi encontrado

  /api/subscritores:
    post:
      tags:
        - Cidadao
      summary: Criar nova conta de Subscritor
      description: Cidadao devera inserir todos os parametros requeridos
      operationId: "criarSubscritor"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Pessoa"
      responses:
        '201':
          description: O pedido foi enviado com sucesso 
        '400':
          description: Algum parametro estava incorreto
        '409': 
          description: O pedido ja se encontra criado.

  /api/subscritores/{nif}:
    get:
      tags:
        - Subscritores
      summary: Pedir informaçao de Subscritor
      operationId: "pedirSubscritor"
      parameters:
      - name: nif
        in: path
        description: nif do cidadao
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Toda a informaçao do Subscritor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscritor'
        '400':
          description: Algum parametro estava incorreto
        '404': 
          description: Recurso nao encontrado

  /api/subscritores/{nif}/veiculos:
    post:
      tags:
        - Subscritores
      summary: Adicionar veiculo ao subscritor
      description: Adicionar novo veiculo ao Subscritor
      operationId: "adicionarVeiculo"
      parameters:
      - name: nif
        in: path
        description: nif do cidadao
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Veiculo"
      responses:
        '201':
          description: O pedido foi enviado com sucesso 
        '400':
          description: Algum parametro estava incorreto
        '409': 
          description: O pedido ja se encontra criado.
    get:
      tags:
        - Subscritores
      summary: Pedir todos os veiculos
      operationId: "pedirVeiculos"
      parameters:
      - name: nif
        in: path
        description: nif do cidadao
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Todos os veiculos do cidadao com sessao aberta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Veiculo'
        '400':
          description: Algum parametro estava incorreto
        '404': 
          description: Recurso nao encontrado

  /api/subscritores/{nif}/veiculos/{matricula}:
    delete:
      tags:
        - Subscritores
      summary: Dissociar o veiculo ao Subscritor
      operationId: "removerVeiculo"
      parameters:
      - name: nif
        in: path
        description: nif do cidadao
        required: true
        schema:
          type: integer
      - in: path
        name: matricula
        required: true
        description: matricula
        schema:
          type: string
      responses:
        '200':
          description: O veiculo foi removido com sucesso
        '202':
          description: Operaçao aceite
        '204':
          description: O veiculo especifico nao se encontra na aplicaçao
        '403':
          description: Nao possui a autorizaçao necessaria

  /api/subscritores/{nif}/notificacoes:
    get:
      tags:
        - Subscritores
      summary: Requerer todas as notificações
      operationId: "pedirNotificações"
      parameters:
      - name: nif
        in: path
        description: nif do cidadao
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Todas as notificações do Subscritor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notificacoes'
        '400':
          description: Algum parametro estava incorreto
        '404': 
          description: Recurso nao encontrado
    post:
      tags:
        - Subscritores
      summary: Adicionar notificacao ao subscritor
      description: Adicionar nova notificacao ao Subscritor
      operationId: "adicionarNotificacao"
      parameters:
      - name: nif
        in: path
        description: nif do cidadao
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventoDeTransito"
      responses:
        '201':
          description: O pedido foi enviado com sucesso 
        '400':
          description: Algum parametro estava incorreto
        '409': 
          description: O pedido ja se encontra criado.

  /api/subscritores/{nif}/notificacoes/{id}:
    get:
      tags:
        - Subscritores
      summary: Requerer a notificaçao com o id
      operationId: "pedirNotificaçao"
      parameters:
      - name: nif
        in: path
        description: nif do cidadao
        required: true
        schema:
          type: integer
      - name: id
        in: path
        description: id da notificaçao
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Todas as notificações do Subscritor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventoDeTransito'
        '400':
          description: Algum parametro estava incorreto
        '404': 
          description: Recurso nao encontrado

  /api/login:
    post:
      tags:
      - Subscritores
      summary: Permite o Subscritor fazer login na aplicaçao
      operationId: loginDeSubscritor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Login"
      responses:
        "200":
          description: operaçao com sucesso
        "400":
          description: Password ou username invalido
          
  /api/logout:
    post:
      tags:
      - Subscritores
      summary: Permite o Subscritor encerrar a sessao
      operationId: logoutDeSubscritor
      responses:
        "200":
          description: operaçao com sucesso
components:
  schemas:
    Subscritor:
      type: object
      properties:
        nif:
          type: integer
        password:
          type: string
        notificações:
          type: array
        veiculos:
          type: array
        veiculosAlugados:
          type: array
        emprestimos:
          type: array
    Admin:
      type: object
      properties:
        username: 
          type: string
        password:
          type: string
    EventoDeTransito:
      type: object
      properties:
        numeroAuto:
          type: string
        veiculoAutuado:
          type: string
        data:
          type: integer
        local:
          type: string
        normaInfrigida:
          type: string
        distrito:
          type: string
        descricaoSumaria:
          type: string
        dataLimiteDefesa:
          type: integer
        estadoDoPagamento:
          type: string
        valorDaCoima:
          type: integer
        gravidade:
          type: string
        entidadeAutuante:
          type: string
        dataNoticacao:
          type: string
    Veiculo:
      type: object
      properties:
        matricula:
          type: string
        modelo: 
          type: string
        categoria: 
          type: string
    Notificacoes:
      type: object
      properties:
        Notificacoes:
          type: array
          items:
            $ref: '#/components/schemas/EventoDeTransito'
    Pagamento:
      type: object
      properties:
        metodoDePagamento: 
          type: string
    PedidoDeDefesa:
      type: object
      properties:
        id:
          type: integer
        descricao:
          type: string
    Emprestimo:
      type: object
      properties:
        dataInicio:
          type: string
        dataFim:
          type: string
        estado:
          type: string
        matricula:
          type: string
    Emprestimos:
      type: object
      properties:
        emprestimos:
          type: array
          items:
            $ref: '#/components/schemas/Emprestimo'
    Login:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    Pessoa:
      type: object
      properties:
        nome:
          type: string
        nif:
          type: string
        numeroConducao:
          type: string
        email:
          type: string